Fixed cancel so it doesn't cancel approved or rejected requests

+++ EmployeeReimburseApp/src/main/java/com/example/employeereimburseapp/CardEmployee.java
@@ -4,6 +4,7 @@ import javafx.fxml.FXML;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class CardEmployee extends RequestCard {
@@ -16,7 +17,16 @@ public class CardEmployee extends RequestCard {
     @FXML
     public void cancelRequest(){
         try(Connection con = CentralDatabase.getConnection()){
-            PreparedStatement ps = con.prepareStatement("DELETE FROM \"Request\" WHERE \"req_id\" = ?");
+            PreparedStatement ps = con.prepareStatement("SELECT \"status\" FROM \"Request\" WHERE \"req_id\" = ?");
+            ps.setInt(1, getRequestID());
+            ResultSet rs = ps.executeQuery();
+            rs.next();
+            String status = rs.getString("status");
+            if(status.equals("Rejected") || status.equals("Approved")){
+                System.out.println("Request cannot be cancelled");
+                return;
+            }
+            ps = con.prepareStatement("DELETE FROM \"Request\" WHERE \"req_id\" = ?");
             ps.setInt(1, getRequestID());
             ps.executeUpdate();
             System.out.println("Request deleted");
